new architecture for socket handlers
